pyproject.toml
[tool.poetry]
name = "cr-kyoushi-generator"
version = "0.0.0"
description = ""
authors = ["Maximilian Frank <maximilian.frank@ait.ac.at>"]
license = "MIT"
readme = 'README.md'  # Markdown files are supported
repository = "https://git-service.ait.ac.at/sct-cyberrange/tools/kyoushi/generator"
homepage = "https://git-service.ait.ac.at/sct-cyberrange/tools/kyoushi/generator"
packages = [
  { include = "cr_kyoushi", from = "src" }
]

[tool.poetry.plugins."cr_kyoushi.generator"]
"random" = "cr_kyoushi.generator.core:RandomGenerator"


[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.8.1"
"ruamel.yaml" = "^0.17.2"
Jinja2 = "^2.11.3"
elasticsearch = "^7.12.0"
elasticsearch-dsl = "^7.3.0"
click = "^7.1.2"
certifi = "^2020.12.5"
pyshark = "^0.4.3"
aiohttp = "^3.7.4"
ujson = "^4.0.2"
livereload = "^2.6.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
pre-commit = "^2.9.0"
black = "^20.8b1"
mypy = "^0.790"
pytest-cov = "^2.10.1"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.5.1"
isort = "^5.7.0"
pylint = "^2.6.0"
safety = "^1.10.3"
autoflake = "^1.4"
mkdocs = "^1.1.2"
mkdocs-material = "^6.1.7"
mkdocs-exclude = "^1.0.2"
mkdocstrings = "^0.13.6"
mkdocs-click = "^0.2.0"
markdown-include = "^0.6.0"
pymdown-extensions = "^8.0.1"
mkdocs-macros-plugin = "^0.5.0"

[tool.isort]
profile = "black"
force_single_line = false
force_grid_wrap = 2
multi_line_output = 3
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
src_paths = ["src", "tests"]
skip_glob = ["*/setup.py"]
filter_files = true
known_first_party = "cr_kyoushi"


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''

[tool.poetry.scripts]
cr-kyoushi-generator = "cr_kyoushi.generator.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
